+++ rust/attest/src/svr2.rs
@@ -5,7 +5,7 @@
 
 use prost::Message;
 
-use crate::constants::{EXPECTED_RAFT_CONFIG_SVR2, SVR2_POSTQUANTUM_OVERRIDE};
+use crate::constants::EXPECTED_RAFT_CONFIG_SVR2;
 use crate::enclave::{Error, Handshake, HandshakeType, Result};
 use crate::proto::svr;
 use crate::util::get_sw_advisories;
@@ -75,10 +75,7 @@ pub fn new_handshake(
         current_time,
         get_sw_advisories(mrenclave),
         expected_raft_config,
-        SVR2_POSTQUANTUM_OVERRIDE
-            .get(&mrenclave)
-            .copied()
-            .unwrap_or(HandshakeType::PostQuantum),
+        HandshakeType::PostQuantum,
     )
 }
 
@@ -116,9 +113,9 @@ mod tests {
     #[test]
     fn attest_svr2() {
         const HANDSHAKE_BYTES: &[u8] = include_bytes!("../tests/data/svr2handshakestart.data");
-        let current_time = SystemTime::UNIX_EPOCH + Duration::from_secs(1741649483);
+        let current_time = SystemTime::UNIX_EPOCH + Duration::from_secs(1768516141);
         let mrenclave_bytes =
-            hex!("38e01eff4fe357dc0b0e8ef7a44b4abc5489fbccba3a78780f3872c277f62bf3");
+            hex!("97f151f6ed078edbbfd72fa9cae694dcc08353f1f5e8d9ccd79a971b10ffc535");
         new_handshake_with_constants(
             &mrenclave_bytes,
             HANDSHAKE_BYTES,
@@ -126,14 +123,14 @@ mod tests {
             &["INTEL-SA-00615", "INTEL-SA-00657"] as &[&str],
             &RaftConfig {
                 min_voting_replicas: 3,
-                max_voting_replicas: 5,
+                max_voting_replicas: 9,
                 super_majority: 0,
-                group_id: 3565209795906488720,
+                group_id: 2330628069874851020,
                 db_version: 2,
                 attestation_timeout: 604800,
                 simulated: false,
             },
-            HandshakeType::PreQuantum,
+            HandshakeType::PostQuantum,
         )
         .unwrap();
     }
@@ -141,26 +138,26 @@ mod tests {
     #[test]
     fn attest_svr2_bad_config() {
         const HANDSHAKE_BYTES: &[u8] = include_bytes!("../tests/data/svr2handshakestart.data");
-        let current_time = SystemTime::UNIX_EPOCH + Duration::from_secs(1741649483);
+        let current_time = SystemTime::UNIX_EPOCH + Duration::from_secs(1768516141);
         let mrenclave_bytes =
-            hex!("38e01eff4fe357dc0b0e8ef7a44b4abc5489fbccba3a78780f3872c277f62bf3");
+            hex!("97f151f6ed078edbbfd72fa9cae694dcc08353f1f5e8d9ccd79a971b10ffc535");
 
         assert!(
             new_handshake_with_constants(
                 &mrenclave_bytes,
                 HANDSHAKE_BYTES,
                 current_time,
-                &[],
+                &["INTEL-SA-00615", "INTEL-SA-00657"] as &[&str],
                 &RaftConfig {
                     min_voting_replicas: 3,
-                    max_voting_replicas: 5,
+                    max_voting_replicas: 9,
                     super_majority: 0,
                     group_id: 0, // wrong
                     db_version: 2,
                     attestation_timeout: 604800,
                     simulated: false,
                 },
-                HandshakeType::PreQuantum,
+                HandshakeType::PostQuantum,
             )
             .is_err()
         );