+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
 v0.86.13
 
 * Limit the number of call link root key bytes used for CallLinkSecretParams derivation.
+
+* updated SVR2 and SVRB enclave configurations